+++ src/cutadapt/seqio.py
@@ -104,52 +104,6 @@ except ImportError:
 	pass
 
 
-class ColorspaceSequence(Sequence):
-	def __init__(self, name, sequence, qualities, primer=None, second_header=False):
-		# In colorspace, the first character is the last nucleotide of the primer base
-		# and the second character encodes the transition from the primer base to the
-		# first real base of the read.
-		if primer is None:
-			self.primer = sequence[0:1]
-			sequence = sequence[1:]
-		else:
-			self.primer = primer
-		if qualities is not None and len(sequence) != len(qualities):
-			rname = _shorten(name)
-			raise FormatError("In read named {!r}: length of colorspace quality "
-				"sequence ({}) and length of read ({}) do not match (primer "
-				"is: {!r})".format(rname, len(qualities), len(sequence), self.primer))
-		super(ColorspaceSequence, self).__init__(name, sequence, qualities, second_header)
-		if self.primer not in ('A', 'C', 'G', 'T'):
-			raise FormatError("Primer base is {!r} in read {!r}, but it "
-				"should be one of A, C, G, T.".format(
-					self.primer, _shorten(name)))
-
-	def __repr__(self):
-		qstr = ''
-		if self.qualities is not None:
-			qstr = ', qualities={!r}'.format(_shorten(self.qualities))
-		return '<ColorspaceSequence(name={!r}, primer={!r}, sequence={!r}{})>'.format(
-			_shorten(self.name), self.primer, _shorten(self.sequence), qstr)
-
-	def __getitem__(self, key):
-		return self.__class__(
-			self.name,
-			self.sequence[key],
-			self.qualities[key] if self.qualities is not None else None,
-			self.primer,
-			self.second_header)
-
-	def __reduce__(self):
-		return (ColorspaceSequence, (self.name, self.sequence, self.qualities, self.primer,
-			self.second_header))
-
-
-def sra_colorspace_sequence(name, sequence, qualities, second_header):
-	"""Factory for an SRA colorspace sequence (which has one quality value too many)"""
-	return ColorspaceSequence(name, sequence, qualities[1:], second_header=second_header)
-
-
 class FileWithPrependedLine:
 	"""
 	A file-like object that allows to "prepend" a single
@@ -225,11 +179,6 @@ class FastaReader(SequenceReader):
 			yield self.sequence_class(name, self._delimiter.join(seq), None)
 
 
-class ColorspaceFastaReader(FastaReader):
-	def __init__(self, file, keep_linebreaks=False):
-		super(ColorspaceFastaReader, self).__init__(file, keep_linebreaks, sequence_class=ColorspaceSequence)
-
-
 class FastqReader(SequenceReader):
 	"""
 	Reader for FASTQ files. Does not support multi-line FASTQ files.
@@ -238,8 +187,7 @@ class FastqReader(SequenceReader):
 		"""
 		file is a path or a file-like object. compressed files are supported.
 
-		The sequence_class should be a class such as Sequence or
-		ColorspaceSequence.
+		The sequence_class should be a class such as Sequence
 		"""
 		super(FastqReader, self).__init__(file)
 		self.sequence_class = sequence_class
@@ -286,16 +234,6 @@ try:
 	from ._seqio import FastqReader
 except ImportError:
 	pass
-
-
-class ColorspaceFastqReader(FastqReader):
-	def __init__(self, file):
-		super(ColorspaceFastqReader, self).__init__(file, sequence_class=ColorspaceSequence)
-
-
-class SRAColorspaceFastqReader(FastqReader):
-	def __init__(self, file):
-		super(SRAColorspaceFastqReader, self).__init__(file, sequence_class=sra_colorspace_sequence)
 
 
 class FastaQualReader:
@@ -348,11 +286,6 @@ class FastaQualReader:
 		self.close()
 
 
-class ColorspaceFastaQualReader(FastaQualReader):
-	def __init__(self, fastafile, qualfile):
-		super(ColorspaceFastaQualReader, self).__init__(fastafile, qualfile, sequence_class=ColorspaceSequence)
-
-
 def sequence_names_match(r1, r2):
 	"""
 	Check whether the sequences r1 and r2 have identical names, ignoring a
@@ -377,9 +310,9 @@ class PairedSequenceReader:
 	"""
 	paired = True
 
-	def __init__(self, file1, file2, colorspace=False, fileformat=None):
-		self.reader1 = open(file1, colorspace=colorspace, fileformat=fileformat)
-		self.reader2 = open(file2, colorspace=colorspace, fileformat=fileformat)
+	def __init__(self, file1, file2, fileformat=None):
+		self.reader1 = open(file1, fileformat=fileformat)
+		self.reader2 = open(file2, fileformat=fileformat)
 		self.delivers_qualities = self.reader1.delivers_qualities
 
 	def __iter__(self):
@@ -427,8 +360,8 @@ class InterleavedSequenceReader:
 	"""
 	paired = True
 
-	def __init__(self, file, colorspace=False, fileformat=None):
-		self.reader = open(file, colorspace=colorspace, fileformat=fileformat)
+	def __init__(self, file, fileformat=None):
+		self.reader = open(file, fileformat=fileformat)
 		self.delivers_qualities = self.reader.delivers_qualities
 
 	def __iter__(self):
@@ -525,13 +458,6 @@ class FastaWriter(FileWriter, SingleRecordWriter):
 			print('>{}'.format(name), sequence, file=self._file, sep='\n')
 
 
-class ColorspaceFastaWriter(FastaWriter):
-	def write(self, record):
-		name = record.name
-		sequence = record.primer + record.sequence
-		super(ColorspaceFastaWriter, self).write(name, sequence)
-
-
 class FastqWriter(FileWriter, SingleRecordWriter):
 	"""
 	Write sequences with qualities in FASTQ format.
@@ -558,14 +484,6 @@ class FastqWriter(FileWriter, SingleRecordWriter):
 			name, sequence, qualities), file=self._file)
 
 
-class ColorspaceFastqWriter(FastqWriter):
-	def write(self, record):
-		name = record.name
-		sequence = record.primer + record.sequence
-		qualities = record.qualities
-		super(ColorspaceFastqWriter, self).writeseq(name, sequence, qualities)
-
-
 class PairRecordWriter:
 	"""Public interface to paired-record files"""
 	def write(self, read1, read2):
@@ -583,10 +501,10 @@ class PairRecordWriter:
 
 
 class PairedSequenceWriter(PairRecordWriter):
-	def __init__(self, file1, file2, colorspace=False, fileformat='fastq', qualities=None):
-		self._writer1 = open(file1, colorspace=colorspace, fileformat=fileformat, mode='w',
+	def __init__(self, file1, file2, fileformat='fastq', qualities=None):
+		self._writer1 = open(file1, fileformat=fileformat, mode='w',
 			qualities=qualities)
-		self._writer2 = open(file2, colorspace=colorspace, fileformat=fileformat, mode='w',
+		self._writer2 = open(file2, fileformat=fileformat, mode='w',
 			qualities=qualities)
 
 	def write(self, read1, read2):
@@ -602,9 +520,9 @@ class InterleavedSequenceWriter(PairRecordWriter):
 	"""
 	Write paired-end reads to an interleaved FASTA or FASTQ file
 	"""
-	def __init__(self, file, colorspace=False, fileformat='fastq', qualities=None):
+	def __init__(self, file, fileformat='fastq', qualities=None):
 		self._writer = open(
-			file, colorspace=colorspace, fileformat=fileformat, mode='w', qualities=qualities)
+			file, fileformat=fileformat, mode='w', qualities=qualities)
 
 	def write(self, read1, read2):
 		self._writer.write(read1)
@@ -621,7 +539,7 @@ class UnknownFileType(Exception):
 
 
 # TODO rename
-def open(file1, file2=None, qualfile=None, colorspace=False, fileformat=None,
+def open(file1, file2=None, qualfile=None, fileformat=None,
 	interleaved=False, mode='r', qualities=None):
 	"""
 	Open sequence files in FASTA or FASTQ format for reading or writing. This is
@@ -639,12 +557,8 @@ def open(file1, file2=None, qualfile=None, colorspace=False, fileformat=None,
 	interleaved -- If True, then file1 contains interleaved paired-end data.
 		file2 and qualfile must be None in this case.
 
-	colorspace -- If True, instances of the Colorspace... classes
-		are returned.
-
 	fileformat -- If set to None, file format is autodetected from the file name
 		extension. Set to 'fasta', 'fastq', or 'sra-fastq' to not auto-detect.
-		Colorspace is not auto-detected and must always be requested explicitly.
 
 	qualities -- When mode is 'w' and fileformat is None, this can be set to
 		True or False to specify whether the written sequences will have quality
@@ -663,28 +577,24 @@ def open(file1, file2=None, qualfile=None, colorspace=False, fileformat=None,
 		raise ValueError("Setting both file2 and qualfile is not supported")
 	if file2 is not None:
 		if mode == 'r':
-			return PairedSequenceReader(file1, file2, colorspace, fileformat)
+			return PairedSequenceReader(file1, file2, fileformat)
 		else:
-			return PairedSequenceWriter(file1, file2, colorspace, fileformat, qualities)
+			return PairedSequenceWriter(file1, file2, fileformat, qualities)
 
 	if interleaved:
 		if mode == 'r':
-			return InterleavedSequenceReader(file1, colorspace, fileformat)
+			return InterleavedSequenceReader(file1, fileformat)
 		else:
-			return InterleavedSequenceWriter(file1, colorspace, fileformat, qualities)
+			return InterleavedSequenceWriter(file1, fileformat, qualities)
 
 	if qualfile is not None:
 		if mode == 'w':
 			raise NotImplementedError('Writing to csfasta/qual not supported')
-		if colorspace:
-			# read from .(CS)FASTA/.QUAL
-			return ColorspaceFastaQualReader(file1, qualfile)
-		else:
-			return FastaQualReader(file1, qualfile)
+		return FastaQualReader(file1, qualfile)
 
 	# All the multi-file things have been dealt with, delegate rest to the
 	# single-file function.
-	return _seqopen1(file1, colorspace=colorspace, fileformat=fileformat,
+	return _seqopen1(file1, fileformat=fileformat,
 		mode=mode, qualities=qualities)
 
 
@@ -707,16 +617,16 @@ def _detect_format_from_name(name):
 	return None
 
 
-def _seqopen1(file, colorspace=False, fileformat=None, mode='r', qualities=None):
+def _seqopen1(file, fileformat=None, mode='r', qualities=None):
 	"""
 	Open a single sequence file. See description above.
 	"""
 	if mode == 'r':
-		fastq_handler = ColorspaceFastqReader if colorspace else FastqReader
-		fasta_handler = ColorspaceFastaReader if colorspace else FastaReader
+		fastq_handler = FastqReader
+		fasta_handler = FastaReader
 	elif mode == 'w':
-		fastq_handler = ColorspaceFastqWriter if colorspace else FastqWriter
-		fasta_handler = ColorspaceFastaWriter if colorspace else FastaWriter
+		fastq_handler = FastqWriter
+		fasta_handler = FastaWriter
 	else:
 		raise ValueError("Mode must be 'r' or 'w'")
 
@@ -726,13 +636,9 @@ def _seqopen1(file, colorspace=False, fileformat=None, mode='r', qualities=None)
 			return fasta_handler(file)
 		elif fileformat == 'fastq':
 			return fastq_handler(file)
-		elif fileformat == 'sra-fastq' and colorspace:
-			if mode == 'w':
-				raise NotImplementedError('Writing to sra-fastq not supported')
-			return SRAColorspaceFastqReader(file)
 		else:
 			raise UnknownFileType("File format {!r} is unknown (expected "
-				"'sra-fastq' (only for colorspace), 'fasta' or 'fastq').".format(fileformat))
+				"'fasta' or 'fastq').".format(fileformat))
 
 	# Detect file format
 	name = None